pyproject.toml
[project]
name = "ftt"
version = "0.1.0"
description = ""
authors = [
    {name = "rukasuleonhart",email = "lucasoliveirasantossoares@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "pytz (>=2025.2,<2026.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]

[tool.pytest.ini_options]
pythonpath = " . "
addopts = ' -p no:warnings'

[tool.ruff]
line-length= 129
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.11.2"
factory-boy = "^3.3.3"

[tool.taskipy.tasks]
run = 'fastapi dev ftt/app.py'
pre_test = 'task lint'
test = 'pytest --cov=ftt -vv -x'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
